% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Add_R2}
\alias{Add_R2}
\title{Add R2 to ggplot}
\usage{
Add_R2(...)
}
\arguments{
\item{...}{Additional params to add to stat_poly_eq.}
}
\value{
A ggplot layer
}
\description{
This is an idea and code taken from
Anatolli Tsyplenkov https://github.com/atsyplenkov to add the R2 statistic to a ggplot.
}
