% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_draw.R
\name{aoi_unzip}
\alias{aoi_unzip}
\title{Title}
\usage{
aoi_unzip(filename, inDir, outDir = tempfile(), pt = F)
}
\arguments{
\item{filename}{A \code{characater}. Input file name, e.g. 'aoi_draw'.}

\item{inDir}{A \code{characater} indicating where .zip file is, e.g. 'D:/downloads'.}

\item{outDir}{A \code{characater} indicating where shape file should go, e.g. 'D:/downloads'.}
}
\description{
Title
}
