% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo.R
\name{cairo_view}
\alias{cairo_view}
\title{Graph Style
This function sets the graphics to 'cairo-png' for better graphing in the IDE. It also sets
the ggplot2 theme to 'theme_light()'.}
\usage{
cairo_view()
}
\value{

}
\description{
Graph Style
This function sets the graphics to 'cairo-png' for better graphing in the IDE. It also sets
the ggplot2 theme to 'theme_light()'.
}
