% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo.R
\name{convert_to_timeframe}
\alias{convert_to_timeframe}
\title{Convert Date and Time}
\usage{
convert_to_timeframe(data, unit)
}
\arguments{
\item{data}{A data.frame with dt variable}

\item{unit}{How to parse the time, e.g. '2 hour', '1 day'.}
}
\value{

}
\description{
Convert Date and Time
}
