% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{discharge_cleaning}
\alias{discharge_cleaning}
\title{Cleaning Data Frame}
\usage{
discharge_cleaning(data, wy_month = 10)
}
\arguments{
\item{data}{A data.frame from \code{get_master_flow()}.}

\item{wy_month}{A numeric vector for water year month.}
}
\value{
A data.frame
}
\description{
This function helps clean the data from the districts
master flow data as well as perform some interpolation for discharge.
}
\examples{
\dontrun{
master_flow <- get_master_flow()
cleaning_update <- discharge_cleaning(master_flow)

}
}
