% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{finalize_q_data}
\alias{finalize_q_data}
\title{Finalize Q Data}
\usage{
finalize_q_data(data, wl_sites_daily, wl_sites_hourly, year, wy_month = 10)
}
\arguments{
\item{data}{Previously created data.frame using \link{get_master_flow}.}

\item{wl_sites_daily}{Previously created list with daily hobo data.}

\item{wl_sites_hourly}{Previously created list with hourly hobo data.}

\item{year}{A numeric vector determining what year to regress the \verb{wl_sites_*} on.}

\item{wy_month}{A numeric vector determining the water year month.}
}
\value{
A data.frame
}
\description{
This function will combine all the previously
created lists (\code{wl_sites_daily} and \code{wl_sites_hourly}) and perform
the log10(Q)~log10(height) regression parameters to daily and hourly data. It will
then add to the original \code{data} for a collated final dataset.
}
