% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal.R
\name{gdal_clip_ogrsf}
\alias{gdal_clip_ogrsf}
\title{Get OGR Clip}
\usage{
gdal_clip_ogrsf(sf, infile, outfile, layers)
}
\arguments{
\item{sf}{An sf object.}

\item{infile}{A file path to the simple feature to clip.}

\item{outfile}{A file path to write the clip to.}

\item{layers}{A character vector. In case you are using a GPKG or gdb you can add the layers here.}
}
\value{

}
\description{
Get OGR (OpenGIS Simple Features Reference Implementation)
Simple Features Library to clip simple features with GDAL. Provide an sf object and clip large
simple feature (infile). The
clip will be based on the bounding box of the sf object.
}
\note{
Sometimes you will need to double quote your path if there are spaces or unique
characters that CLI needs quoted, e.g. '"C:/josh erickson/some file path/"'.
}
