% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal.R
\name{gdal_clip_shape}
\alias{gdal_clip_shape}
\title{Get Shape Clip}
\usage{
gdal_clip_shape(sf, infile, outfile, layers)
}
\arguments{
\item{layers}{}
}
\value{

}
\description{
Get vector clip with GDAL. Provide sf object and clip large vector.
}
