% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils.R
\name{get_compMax_intersection}
\alias{get_compMax_intersection}
\title{Get Max Composition}
\usage{
get_compMax_intersection(soildf, sf)
}
\arguments{
\item{soildf}{a previously created get_HSG() or get_soilDB_sf() object.}

\item{sf}{a sf object that intersects soildf}
}
\description{
Get Max Composition
}
