% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils.R
\name{get_HSG}
\alias{get_HSG}
\title{Get HSG}
\usage{
get_HSG(sf)
}
\arguments{
\item{sf}{}
}
\value{

}
\description{
Gets hydrologic soil group per composition per polygon
but also a lot of other ssurgo information.
}
