% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_hydro_appendix}
\alias{get_hydro_appendix}
\title{Get Hydro Appendix}
\usage{
get_hydro_appendix(
  save_path,
  eval_plot = TRUE,
  eval_huc_table = TRUE,
  eval_prop_table = TRUE,
  eval_roads_table = TRUE,
  ...
)
}
\arguments{
\item{save_path}{A file path to save the .docx file.}

\item{eval_plot}{Render a TSS plot.}

\item{eval_huc_table}{Render a HUC table.}

\item{eval_prop_table}{Render a proportion table.}

\item{...}{Arguments to pass to \link{get_huc12_information} and \link{get_treatment_proportion}.}
}
\value{
A rendered HTML report.
}
\description{
Get Hydro Appendix
}
