% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils.R
\name{get_soil_proportion}
\alias{get_soil_proportion}
\title{Get Soil Proportion}
\usage{
get_soil_proportion(soil_catch)
}
\arguments{
\item{soil_catch}{a previously created get_compMax_intersection() object}
}
\description{
Get Soil Proportion
}
