% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_soils_appendix}
\alias{get_soils_appendix}
\title{Get Soils Appendix}
\usage{
get_soils_appendix(path, save_path, ...)
}
\arguments{
\item{path}{A file path to the polygons.}

\item{save_path}{A file path to save the .docx file.}

\item{...}{If no path, then ... can be an sf object.}
}
\value{
A rendered HTML report
}
\description{
Get Soils Appendix
}
