% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{get_treatment_proportion}
\alias{get_treatment_proportion}
\title{Getting Treatment Proportions}
\usage{
get_treatment_proportion(path = NULL, sf = NULL, ...)
}
\arguments{
\item{path}{character; typically the T:Drive}

\item{sf}{An sf object.}

\item{...}{Additional arguments.}
}
\value{
A data.frame
}
\description{
Getting Treatment Proportions
}
