% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{get_tss_timeseries}
\alias{get_tss_timeseries}
\title{Get TSS Timeseries}
\usage{
get_tss_timeseries(stream, year, ...)
}
\arguments{
\item{stream}{A character vector of stream names}

\item{year}{A numeric vector of years, e.g. 2012:2015, 2015, etc.}

\item{...}{Arguments to pass to appedix functions.}
}
\value{
A ggplot.
}
\description{
Get TSS Timeseries
}
