% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{get_wy_stats}
\alias{get_wy_stats}
\title{Water Year Stats}
\usage{
get_wy_stats(procDV)
}
\arguments{
\item{procDV}{A previously created \link[resourceviz]{discharge_cleaning}.}
}
\value{
A \code{tibble} filtered by water year with added meta-data.
}
\description{
This function uses the results of the \link[resourceviz]{discharge_cleaning} object to
generate mean, maximum, median, standard deviation and some normalization methods (drainage
area, scaled by log and standard deviation) per water year.
}
\examples{
\dontrun{

library(resourceviz)

master_flow <- get_master_flow()
cleaning_update <- discharge_cleaning(master_flow)

wy_stats <- get_wy_stats(cleaning_update)
}

}
