% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap.R
\name{google_map}
\alias{google_map}
\title{Title}
\usage{
google_map(aoi, maptype = "terrain", source = "google", zoom = 12)
}
\arguments{
\item{aoi}{Area of Interest}

\item{maptype}{What google maptype to use,}

\item{zoom}{}
}
\description{
Title
}
\note{
google map types: "terrain", "terrain-background", "satellite", "roadmap", "hybrid",
"toner", "watercolor", "terrain-labels", "terrain-lines", "toner-2010", "toner-2011",
"toner-background", "toner-hybrid", "toner-labels", "toner-lines", "toner-lite".
}
