% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hydro_huc_table}
\alias{hydro_huc_table}
\title{Generate gt Hydro Table}
\usage{
hydro_huc_table(data)
}
\arguments{
\item{data}{A previously created \link{get_huc12_information} object.}
}
\value{
A gt table.
}
\description{
Generate gt Hydro Table
}
