% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo.R
\name{light_and_cairo}
\alias{light_and_cairo}
\title{Graph Style
This function sets the graphics to 'cairo-png' for better graphing in the IDE. It also sets
the ggplot2 theme to 'theme_light()'.}
\usage{
light_and_cairo()
}
\value{

}
\description{
Graph Style
This function sets the graphics to 'cairo-png' for better graphing in the IDE. It also sets
the ggplot2 theme to 'theme_light()'.
}
