% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo.R
\name{read_hobo_csv}
\alias{read_hobo_csv}
\title{Read-in HOBO files}
\usage{
read_hobo_csv(hobo_csv = NULL, pressure_unit)
}
\arguments{
\item{hobo_csv}{Either a folder path or path to csv.}

\item{pressure_unit}{A \code{character} of input pressure unit}
}
\value{
a tibble with HOBO attributes: logger_sn, dt, date, time,abs_pres, temp_f
}
\description{
This function takes in a previously created csv file or multiple csv files within
the HOBOware software.
}
