% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sites_to_list}
\alias{sites_to_list}
\title{Helper List Function}
\usage{
sites_to_list(names = NULL)
}
\arguments{
\item{names}{A character vector indicating names you want discarded in list.}
}
\value{
A list
}
\description{
This helps get the stations with the transducers into a list
after they've been processed.
}
