% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_USGS_box}
\alias{theme_USGS_box}
\title{USGS Theme}
\usage{
theme_USGS_box(base_family = "serif", ...)
}
\arguments{
\item{base_family}{character}

\item{...}{Arguments to pass on to theme_bw().}
}
\value{
A ggplot theme layer.
}
\description{
USGS Theme
}
