% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{treatment_table}
\alias{treatment_table}
\title{Get Treatment Table}
\usage{
treatment_table(data)
}
\arguments{
\item{data}{A previously created \link{get_treatment_propotion} object.}
}
\description{
Get Treatment Table
}
