% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NLDI.R
\name{viz_GridMET}
\alias{viz_GridMET}
\title{Plot GridMET}
\usage{
viz_GridMET(
  data,
  row,
  param = "prcp",
  startDate = "2019-01-01",
  endDate = "2020-01-01"
)
}
\arguments{
\item{data}{A sf point object}

\item{row}{\code{numeric}. If a specific place is wanted then enter row.}

\item{param}{\code{character}. See \link[climateR]{getGridMET} for more info.}

\item{startDate}{\code{character}. "2020-01-01"}

\item{endDate}{\code{character}. "2020-01-01"}
}
\value{
A ggplot.
}
\description{
A wrapper function around \link[climateR]{getGridMET} that plot's time series for
selected parameters.
}
