% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NLDI.R
\name{viz_NLDI}
\alias{viz_NLDI}
\title{Title}
\usage{
viz_NLDI(
  sf_pt,
  leaflet = TRUE,
  maptype = "terrain",
  source = "google",
  zoom = 8
)
}
\arguments{
\item{sf_pt}{A sf point object}

\item{leaflet}{TRUE/FALSE for plotting with leaflet}

\item{maptype}{A \code{character} indicating maptype. see \link[ggmap]{get_map} documentation}

\item{zoom}{A \code{numeric} value between 1-13, zoom level for ggmap.}
}
\value{

}
\description{
Title
}
