% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NLDI.R
\name{viz_NLDIcatch}
\alias{viz_NLDIcatch}
\title{Visualise NLDI catchments}
\usage{
viz_NLDIcatch(comidID, return.sf = FALSE)
}
\arguments{
\item{comidID}{A nhdplus comid.}
}
\value{
A leaflet map with select catchment variables.
}
\description{
Visualise NLDI catchments
}
