% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_poly.R
\name{viz_Roads}
\alias{viz_Roads}
\title{Leaflet with INFRA Roads}
\usage{
viz_Roads(data)
}
\arguments{
\item{data}{A INFRA road layer. Variable must be all caps.}
}
\value{
A leafet map with roads colored by Route Status and popup with select info.
}
\description{
Leaflet with INFRA Roads
}
