% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo.R
\name{water_level_conversion}
\alias{water_level_conversion}
\title{Convert to Inches}
\usage{
water_level_conversion(
  baro_df,
  water_df,
  baro_elev,
  hobo_elev,
  gage_height = NULL,
  gh_date = NULL,
  var_select = c("abs_pres_water", "abs_pres_baro", "GH", "gage_height"),
  method = "lasso.select.norm"
)
}
\arguments{
\item{baro_df}{A data.frame with column abs_pres and dt.}

\item{water_df}{A data.frame with column abs_pres and dt.}

\item{baro_elev}{Elevation in feet at the barometer site.}

\item{hobo_elev}{Elevation in feet at the gage site.}

\item{gh_date}{A character vector for the date of \code{add_gh}, e.g. '2022-05-14 11:00:00'.}

\item{var_select}{A character vector indicating what variables to use with \code{mice::mice()} imputation.}

\item{method}{A character indicating what method to use in \code{mice::mice()}, 'lasso.select.norm' (default).}

\item{gh}{A numeric vector indicating the gage height for a particular day(s).}
}
\description{
Convert to Inches
}
\note{
The \code{gh} and \code{gh_date} parameters will perform differently if the
vector length 1, e.g. one gh_date and gh. It will just subtract or add to
the height at that date going forward; however, if greater than 1 then \code{gh} and \code{gh_date}
will be used to try and 'calibrate'
the height of the transducer to a 'observed' gage height \code{height_imp} using \code{mice::mice(method = 'lasso.norm')}.
}
